сделал заглушку на удаление комнаты, если там находятся пользователи

+++ chat_app/tornadoapp.py
@@ -8,14 +8,14 @@ import logging
 import re
 import json
 from copy import copy
-from .models import Room, Message, CustomUser, NotAllowedToChange, NotAllowedToDelete
+from .models import Room, Message, CustomUser, NotAllowedToChange, NotAllowedToDelete, ReachMaxRoomCount
 from django.contrib.sessions.models import Session
 from django.core.exceptions import ObjectDoesNotExist
 from django.db import IntegrityError
 
 
 logging_handler = logging.FileHandler(filename='chat_debug_log.txt', mode='a')
-logging_formatter = logging.Formatter(fmt='[%(levelname)-8s] %(message)s')
+logging_formatter = logging.Formatter(fmt='[%(levelname)-8s %(asctime)s] %(message)s', datefmt='%d-%m-%Y %H:%M:%S')
 logging_handler.setFormatter(logging_formatter)
 logger = logging.getLogger()
 logger.addHandler(logging_handler)
@@ -58,9 +58,9 @@ class ChatSocketHandler(tornado.websocket.WebSocketHandler):
         mess = json.dumps(d)
         self.write_message(mess)  # шлю себе все комнаты
 
-        users_in_m_room = [con.user_name for con in self.connections[self.main_room]]
+        users_in_main_room = [conn.user_name for conn in self.connections[self.main_room]]
         d = {'type': 'all_users',
-             'users': users_in_m_room}
+             'users': users_in_main_room}
         mess = json.dumps(d)
         self.write_message(mess)  # шлю всех польз-й из основной комнатыф
 
@@ -102,6 +102,7 @@ class ChatSocketHandler(tornado.websocket.WebSocketHandler):
             self.close()
         elif message['type'] == 'change_room':
             self.connections[message['new_room']].append(self)
+            self.connections[message['room']].remove(self)
 
             users_in_new_room = [conn.user_name for conn in self.connections[message['new_room']]]
             d = {'type': 'all_users',  # шлю себе всех пользователей из новой комнаты
@@ -112,12 +113,11 @@ class ChatSocketHandler(tornado.websocket.WebSocketHandler):
             for conn in self.connections[message['new_room']]:
                 if conn.user_name != self.user_name:
                     d = {'type': 'add_user',
-                         'room': self.main_room,
+                         'room': message['new_room'],
                          'user': self.user_name}
                     mess = json.dumps(d)
                     conn.write_message(mess)
 
-            self.connections[message['room']].remove(self)
             d = {'type': 'remove_user',
                  'user': self.user_name,
                  'room': message['room']}
@@ -134,6 +134,9 @@ class ChatSocketHandler(tornado.websocket.WebSocketHandler):
             except IntegrityError:
                 d = {'type': 'error',
                      'text': 'Такая комната уже существует'}
+            except ReachMaxRoomCount:
+                d = {'type': 'error',
+                     'text': 'Создано максимально допустимое количество комнат'}
             mess = json.dumps(d)
             all_connections = [conn for room in self.connections for conn in self.connections[room]]
             for conn in all_connections:
@@ -160,16 +163,24 @@ class ChatSocketHandler(tornado.websocket.WebSocketHandler):
                 conn.write_message(mess)
         elif message['type'] == 'delete_room':
             try:
-                room = Room.objects.get(title=message['deleted_room'])
-                room.delete()
-                for conn in self.connections[message['deleted_room']]:
-                    d = {'type': 'drop_to_main',
-                         'user': conn.user_name}
-                    mess = json.dumps(d)
-                    conn.write_message(d)
-                del self.connections[message['deleted_room']]
-                d = {'type': 'delete_room',
-                     'deleted_room': message['deleted_room']}
+                if not self.connections[message['deleted_room']]:
+                    room = Room.objects.get(title=message['deleted_room'])
+                    room.delete()
+                    #users_in_main_room = [conn.user_name for conn in self.connections[self.main_room]]
+                    #users = users_in_main_room + [conn.user_name for conn in self.connections[message['deleted_room']]]
+                    #for conn in self.connections[message['deleted_room']]:
+                    #    d = {'type': 'drop_to_main',
+                    #         'users': users,
+                    #         'room': self.main_room}
+                    #    mess = json.dumps(d)
+                    #    conn.write_message(d)
+                    #self.connections[self.main_room].extend(copy(self.connections[message['deleted_room']]))
+                    del self.connections[message['deleted_room']]
+                    d = {'type': 'delete_room',
+                         'deleted_room': message['deleted_room']}
+                else:
+                    d = {'type': 'error',
+                         'text': 'В комнате еще находятся пользователи'}
             except ObjectDoesNotExist:
                 d = {'type': 'error',
                      'text': 'Такой комнаты не существует'}
@@ -177,6 +188,9 @@ class ChatSocketHandler(tornado.websocket.WebSocketHandler):
                 d = {'type': 'error',
                      'text': 'Нельзя удалить'}
             mess = json.dumps(d)
-            all_connections = [conn for room in self.connections for conn in self.connections[room]]
-            for conn in all_connections:
-                conn.write_message(mess)
+            if d['type'] == 'error':
+                self.write_message(mess)
+            else:
+                all_connections = [conn for room in self.connections for conn in self.connections[room]]
+                for conn in all_connections:
+                    conn.write_message(mess)
